Move client-side entity registries to client.registry package.
