+++ station-entities-v0/src/main/java/net/modificationstation/stationapi/api/event/registry/EntityHandlerRegistryEvent.java
@@ -1,6 +1,6 @@
 package net.modificationstation.stationapi.api.event.registry;
 
-import net.modificationstation.stationapi.api.registry.EntityHandlerRegistry;
+import net.modificationstation.stationapi.api.client.registry.EntityHandlerRegistry;
 
 public class EntityHandlerRegistryEvent extends RegistryEvent<EntityHandlerRegistry> {
     public EntityHandlerRegistryEvent() {
