+++ station-entities-v0/src/main/java/net/modificationstation/stationapi/api/event/registry/MobHandlerRegistryEvent.java
@@ -1,6 +1,6 @@
 package net.modificationstation.stationapi.api.event.registry;
 
-import net.modificationstation.stationapi.api.registry.MobHandlerRegistry;
+import net.modificationstation.stationapi.api.client.registry.MobHandlerRegistry;
 
 public class MobHandlerRegistryEvent extends RegistryEvent<MobHandlerRegistry> {
     public MobHandlerRegistryEvent() {
