+++ station-entities-v0/src/main/java/net/modificationstation/stationapi/mixin/entity/client/ClientNetworkHandlerMixin.java
@@ -4,9 +4,8 @@ import net.minecraft.class_454;
 import net.minecraft.client.network.ClientNetworkHandler;
 import net.minecraft.entity.Entity;
 import net.minecraft.network.packet.s2c.play.EntitySpawnS2CPacket;
-import net.minecraft.world.World;
 import net.modificationstation.stationapi.api.client.entity.factory.EntityWorldAndPosFactory;
-import net.modificationstation.stationapi.api.registry.EntityHandlerRegistry;
+import net.modificationstation.stationapi.api.client.registry.EntityHandlerRegistry;
 import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
